apply suggestistions
