+++ src/jllvm/materialization/ByteCodeCompileLayer.cpp
@@ -24,23 +24,8 @@ void jllvm::ByteCodeCompileLayer::emit(std::unique_ptr<llvm::orc::Materializatio
                                llvm::GlobalValue::ExternalLinkage, mangleMethod(*methodInfo, *classFile), module.get());
     function->setGC("coreclr");
 
-    std::string sectionName = "java";
-    if (llvm::Triple(LLVM_HOST_TRIPLE).isOSBinFormatMachO())
-    {
-        sectionName = "__TEXT," + sectionName;
-        sectionName += ",regular,pure_instructions";
-    }
+    applyJavaMethodAttributes(function, {classObject, method});
 
-    auto* ptrType = llvm::PointerType::get(*context, 0);
-    function->setPrefixData(llvm::ConstantStruct::get(
-        llvm::StructType::get(*context, {referenceType(*context), ptrType}),
-        {llvm::ConstantExpr::getIntToPtr(llvm::ConstantInt::get(m_dataLayout.getIntPtrType(*context),
-                                                                reinterpret_cast<std::uintptr_t>(classObject)),
-                                         referenceType(*context)),
-         llvm::ConstantExpr::getIntToPtr(
-             llvm::ConstantInt::get(m_dataLayout.getIntPtrType(*context), reinterpret_cast<std::uintptr_t>(method)),
-             ptrType)}));
-    function->setSection(sectionName);
     function->addFnAttr(llvm::Attribute::UWTable);
 #ifdef LLVM_ADDRESS_SANITIZER_BUILD
     function->addFnAttr(llvm::Attribute::SanitizeAddress);
